refactor(disease-outbreaks): extract helpers + inject nowMs to kill test drift and timing flake
Greptile P2s on PR #3597:

1. tests/disease-outbreaks-seed.test.mjs replicated parser/mapper/contentMeta
   logic locally — a drift in fetchWhoDonApi or contentMeta would not have
   failed any of the 16 tests because they asserted against their own copy
   of the logic, not the seeder's.

2. The "near-future ≤1h accepted" test relied on Date.now() being stable
   between test setup and the call into contentMeta. On a loaded CI runner
   the gap could exceed the (1h - 30min) margin and flake.

Fixes both at once:

- New scripts/_disease-outbreaks-helpers.mjs exports the pure functions
  (whoNormalizeItem, rssNormalizeItem, tghNormalizeItem, mapItem,
  diseaseContentMeta, diseasePublishTransform, DISEASE_MAX_CONTENT_AGE_MIN).
  diseaseContentMeta accepts an optional nowMs for deterministic skew tests.

- Seeder imports those helpers instead of inlining them. ~150 lines
  removed; behavior unchanged (verified by node -c + smoke test).

- Test file imports the real helpers (no replicas). All skew-limit tests
  inject FIXED_NOW=1700000000000 — no wall-clock dependence.

- Tightens the "within 1h tolerance" test from +30min to +5min ahead of
  injected NOW, well clear of the 1h boundary regardless of the timing fix.

Net: -265 lines across the two existing files; +200 in the new helpers
module. 17/17 disease tests pass; 49/49 across the full Sprint 1+2 stack.
